+++ src/main/java/com/teamtea/eclipticseasons/EclipticSeasons.java
@@ -111,6 +111,7 @@ public class EclipticSeasons {
         BlockRegistry.BLOCK_DEFERRED_REGISTER.register(modEventBus);
         ItemRegistry.ITEM_DEFERRED_REGISTER.register(modEventBus);
         BlockEntityRegistry.BLOCK_ENTITY_TYPE_DEFERRED_REGISTER.register(modEventBus);
+        LootItemConditionRegistry.LOOT_ITEM_CONDITION_TYPE_DEFERRED_REGISTER.register(modEventBus);
 
         modEventBus.addListener(this::gatherData);
         modEventBus.addListener(this::FMLCommonSetup);