+++ src/main/java/com/teamtea/eclipticseasons/data/extend/example/ExampleGiftLootTables.java
@@ -1,5 +1,8 @@
-package com.teamtea.eclipticseasons.data.general.loot;
+package com.teamtea.eclipticseasons.data.extend.example;
 
+import com.teamtea.eclipticseasons.api.constant.solar.Season;
+import com.teamtea.eclipticseasons.api.constant.solar.SolarTerm;
+import com.teamtea.eclipticseasons.common.loot.SeasonCondition;
 import com.teamtea.eclipticseasons.common.registry.ESLootTables;
 import com.teamtea.eclipticseasons.common.registry.ItemRegistry;
 import net.minecraft.data.loot.LootTableSubProvider;
@@ -11,7 +14,7 @@ import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
 
 import java.util.function.BiConsumer;
 
-public class EclipticSeasonsGiftLootTables implements LootTableSubProvider{
+public class ExampleGiftLootTables implements LootTableSubProvider{
 
 
     @Override
@@ -21,19 +24,9 @@ public class EclipticSeasonsGiftLootTables implements LootTableSubProvider{
                 LootTable.lootTable()
                         .withPool(
                                 LootPool.lootPool()
-                                        // .when(SeasonCondition.builder(SeasonCondition.Slice.builder().season(Season.SPRING).build()))
+                                        .when(SeasonCondition.builder(SeasonCondition.Slice.builder().solarTerm(SolarTerm.BEGINNING_OF_SPRING).build()))
                                         .setRolls(ConstantValue.exactly(1.0F))
                                         .add(LootItem.lootTableItem(ItemRegistry.spring_greenhouse_essence_item.get()).setWeight(10))
-                        )
-        );
-
-        output.accept(
-                ESLootTables.summer_greenhouse_essence,
-                LootTable.lootTable()
-                        .withPool(
-                                LootPool.lootPool()
-                                        .setRolls(ConstantValue.exactly(1.0F))
-                                        .add(LootItem.lootTableItem(ItemRegistry.summer_greenhouse_essence_item.get()).setWeight(10))
                         )
         );
 
@@ -42,18 +35,9 @@ public class EclipticSeasonsGiftLootTables implements LootTableSubProvider{
                 LootTable.lootTable()
                         .withPool(
                                 LootPool.lootPool()
+                                        .when(SeasonCondition.builder(SeasonCondition.Slice.builder().start(SolarTerm.AUTUMNAL_EQUINOX).end(SolarTerm.BEGINNING_OF_WINTER).build()))
                                         .setRolls(ConstantValue.exactly(1.0F))
-                                        .add(LootItem.lootTableItem(ItemRegistry.autumn_greenhouse_essence_item.get()).setWeight(10))
-                        )
-        );
-
-        output.accept(
-                ESLootTables.winter_greenhouse_essence,
-                LootTable.lootTable()
-                        .withPool(
-                                LootPool.lootPool()
-                                        .setRolls(ConstantValue.exactly(1.0F))
-                                        .add(LootItem.lootTableItem(ItemRegistry.winter_greenhouse_essence_item.get()).setWeight(10))
+                                        .add(LootItem.lootTableItem(ItemRegistry.spring_greenhouse_essence_item.get()).setWeight(10))
                         )
         );