+++ changelog.md
@@ -1,3 +1,15 @@
+### 0.11.8.4
+
+- Add a control condition named `eclipticseasons:season` for loot, which can be used to specify loot tables by season.
+```json
+{
+  "condition": "eclipticseasons:season",
+  "require": {
+    "start": "autumnal_equinox",
+    "end": "beginning_of_winter"
+  }
+}
+  ```
 ### 0.11.8.3
 
 - Improved biome ID deserialization to reduce null returns.